Move colors, match and test to their own file.
